Add auto send to payload generator
